services/recommendation-ml/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
psycopg2-binary==2.9.10
pgvector==0.3.6
redis==5.2.1
confluent-kafka==2.6.1
scikit-learn==1.6.1
numpy==2.2.2
scipy==1.15.1
python-dotenv==1.0.1
apscheduler==3.11.0
